reorganize fake API files
