+++ .gitignore
@@ -10,4 +10,4 @@ gulp-cache
 /tmp/
 /docker/html/
 package-lock.json
-test/fake-api-db/*.json
+test/fake-api/db/*.json
