+++ cypress/plugins/index.js
@@ -11,8 +11,8 @@
 // This function is called when a project is opened or re-opened (e.g. due to
 // the project's config changing)
 
-const fakeApiDb = require('../../test/fake-api-db');
-const fakeApiToken = require('../../test/fake-api-token');
+const fakeApiDb = require('../../test/fake-api/db');
+const fakeApiToken = require('../../test/fake-api/token');
 
 module.exports = (on) => {
   on('task', {
