+++ test/fake-api/fixtures/granules/MOD09GQ.A2016358.h13v04.006.2016360104606/index.json
