+++ test/fake-api/fixtures/providers/s3_provider/index.json
