+++ test/fake-api/fixtures/rules/MOD09GQ_TEST_kinesisRule/index.json
