+++ test/fake-api/fixtures/schemas/collection/index.json
