+++ test/fake-api/fixtures/schemas/rule/index.json
