+++ test/fake-api/fixtures/stats/aggregate/granules/index.json
